Save work

+++ src/util/tsc.h
@@ -2,34 +2,35 @@
 #define ERPC_TSC_H
 
 #include <stdint.h>
+#include <stdlib.h>
 
 namespace ERpc {
 
 /// Return the TSC
-static inline uint64_t rdtsc() {
+static inline size_t rdtsc() {
   uint64_t rax;
   uint64_t rdx;
   asm volatile("rdtsc" : "=a"(rax), "=d"(rdx));
-  return (rdx << 32) | rax;
+  return static_cast<size_t>((rdx << 32) | rax);
 }
 
 /// Convert cycles measured by rdtsc with frequence \p freq_ghz to seconds
-static double to_sec(uint64_t cycles, double freq_ghz) {
+static double to_sec(size_t cycles, double freq_ghz) {
   return (cycles / (freq_ghz * 1000000000));
 }
 
 /// Convert cycles measured by rdtsc with frequence \p freq_ghz to msec
-static double to_msec(uint64_t cycles, double freq_ghz) {
+static double to_msec(size_t cycles, double freq_ghz) {
   return (cycles / (freq_ghz * 1000000));
 }
 
 /// Convert cycles measured by rdtsc with frequence \p freq_ghz to usec
-static double to_usec(uint64_t cycles, double freq_ghz) {
+static double to_usec(size_t cycles, double freq_ghz) {
   return (cycles / (freq_ghz * 1000));
 }
 
 /// Convert cycles measured by rdtsc with frequence \p freq_ghz to nsec
-static double to_nsec(uint64_t cycles, double freq_ghz) {
+static double to_nsec(size_t cycles, double freq_ghz) {
   return (cycles / freq_ghz);
 }
